Cleanup and updated readme
