+++ cmd/proto_handler/task.go
@@ -7,8 +7,6 @@ import (
 	"github.com/kozloz/togo/internal/errors"
 	"github.com/kozloz/togo/internal/genproto"
 	"github.com/kozloz/togo/internal/tasks"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -17,25 +15,12 @@ type TaskHandler struct {
 	op *tasks.Operation
 }
 
-func Validate(ctx context.Context, createTaskReq *genproto.CreateTaskRequest) error {
-	log.Printf("Validating request: %v", createTaskReq)
-	if createTaskReq.GetUserID() == 0 {
-		log.Printf("Error: Invalid UserID provided")
-		_ = grpc.SetHeader(ctx, metadata.Pairs("x-http-code", "400"))
-		return errors.InvalidUserID
-	}
-	if createTaskReq.GetName() == "" {
-		log.Printf("Error: Empty task name provided")
-		_ = grpc.SetHeader(ctx, metadata.Pairs("x-http-code", "400"))
-		return errors.InvalidTaskName
-	}
-	return nil
-}
-
 func (t *TaskHandler) CreateTask(ctx context.Context, createTaskReq *genproto.CreateTaskRequest) (*genproto.CreateTaskResponse, error) {
+	// Print out the protobuf
 	log.Println(proto.Marshal(createTaskReq))
 
 	createTaskResp := &genproto.CreateTaskResponse{}
+
 	// Validate the request
 	err := Validate(ctx, createTaskReq)
 	if err != nil {
